+++ tests/cameras/mock_cv2.py
