+++ tests/cameras/mock_pyrealsense2.py
