+++ tests/examples/test_examples.py
